Trabajando detalles de Proveedor

+++ src/main/java/cr/ac/una/proyecto_progra4/services/ProveedoresServices.java
@@ -5,7 +5,9 @@
 package cr.ac.una.proyecto_progra4.services;
 
 import cr.ac.una.proyecto_progra4.data.ProveedoresData;
+import cr.ac.una.proyecto_progra4.domain.OrdenDeCompra;
 import cr.ac.una.proyecto_progra4.domain.Proveedor;
+import cr.ac.una.proyecto_progra4.jpa.OrdenDeCompraRepository;
 import cr.ac.una.proyecto_progra4.jpa.ProveedorRepository;
 import java.util.LinkedList;
 import java.util.List;
@@ -52,6 +54,9 @@ public class ProveedoresServices implements IProveedoresService {
 
     @Autowired
     private ProveedorRepository proveedorRep;
+    
+    @Autowired
+    private OrdenDeCompraRepository ordenDeCompraRep;
 
     @Override
     public String guardar(Proveedor proveedor) {
@@ -71,6 +76,12 @@ public class ProveedoresServices implements IProveedoresService {
     @Override
     public String eliminar(int id) {
         try {
+            // Verificar si el proveedor está asociado a alguna orden de compra
+            List<OrdenDeCompra> ordenesDeCompra = ordenDeCompraRep.findByProveedorId(id);
+            if (!ordenesDeCompra.isEmpty()) {
+                return "{\"success\": false, \"message\": \"No se puede eliminar el proveedor porque está asociado a una o más órdenes de compra.\"}";
+            }
+
             proveedorRep.deleteById(id);
             return "{\"success\": true, \"message\": \"¡Proveedor eliminado exitosamente!\"}";
         } catch (Exception e) {
